added options bar for exporting function

+++ app/src/main/java/com/michaelc/contacttracing/ItemAdapter.kt
@@ -1,5 +1,6 @@
 package com.michaelc.contacttracing
 
+import android.app.Activity
 import android.app.AlertDialog
 import android.content.Context
 import android.database.sqlite.SQLiteFullException
@@ -11,8 +12,10 @@ import android.widget.BaseAdapter
 import android.widget.ImageView
 import android.widget.TextView
 import android.widget.Toast
+import com.google.android.material.floatingactionbutton.FloatingActionButton
+import com.google.android.material.snackbar.Snackbar
 import kotlinx.android.synthetic.main.edit_dialog.view.*
-import java.text.DateFormat
+import java.lang.NullPointerException
 import java.text.SimpleDateFormat
 import java.util.*
 
@@ -47,11 +50,21 @@ class ItemAdapter(
 
         val rowView = inflater.inflate(R.layout.item_row, parent, false)
         // Get the name of an element on the page
-        val nameTextView = rowView.findViewById(R.id.tvName) as TextView
-// get the image views for edit and delete icons
-        val editIcon = rowView.findViewById(R.id.ivEdit) as ImageView
-        val deleteIcon = rowView.findViewById(R.id.ivDelete) as ImageView
-      //  val confirmIcon = rowView.findViewById(R.id.mDialogView.dialogConfirmBtn)
+        val nameTextView =  rowView.findViewById(R.id.tvName) as TextView
+        // get the image views for edit and delete icons
+        val editIcon =      rowView.findViewById(R.id.ivEdit) as ImageView
+        val deleteIcon =    rowView.findViewById(R.id.ivDelete) as ImageView
+        // val confirmIcon = rowView.findViewById(R.id.mDialogView.dialogConfirmBtn)
+
+        //tried fab but not suitable
+        /*try {
+            val fab =           rowView.findViewById(R.id.fab) as? FloatingActionButton//floating action button
+
+        }*/
+
+
+
+
 
         editIcon.setOnClickListener {
             //Inflate the dialog with custom view
@@ -68,7 +81,6 @@ class ItemAdapter(
             var time: String
             var date: Date
 
-
             name = a.name
             number = a.number
             time = a.time
@@ -77,8 +89,6 @@ class ItemAdapter(
             val pattern = "dd/MM/yyyy"
             val simpleDateFormat = SimpleDateFormat(pattern)
             val dateString = simpleDateFormat.format(date)
-
-
 
             mDialogView.dialogNameEt.setText(name.toString())
             mDialogView.dialogNumberEt.setText(number.toString())
@@ -124,19 +134,33 @@ class ItemAdapter(
                 mAlertDialog.dismiss()
                 //get text from EditTexts of custom layout
             }
-
         }
 
         deleteIcon.setOnClickListener {
 
-            var a = getItem(position) as ContactDetails
-            var db = DatabaseHandler(context)
-            db.deleteData(a)
+            val mAlertDialog = AlertDialog.Builder(context)
+            mAlertDialog.setIcon(R.mipmap.ic_launcher) //set alertdialog icon
+            mAlertDialog.setTitle("DELETE!") //set alertdialog title
+            mAlertDialog.setMessage("Are you sure you want to delete the contact") //set alertdialog message
+            mAlertDialog.setPositiveButton("Yes") { dialog, id ->
+                //perform some tasks here
+                var a = getItem(position) as ContactDetails
+                var db = DatabaseHandler(context)
+                db.deleteData(a)
 
 
-          val gVData = MainActivity.GlobalVariable.mutableListData
-            gVData.remove(a)
-            notifyDataSetChanged()
+                val gVData = MainActivity.GlobalVariable.mutableListData
+                gVData.remove(a)
+                notifyDataSetChanged()
+                Toast.makeText(context, "Yes", Toast.LENGTH_SHORT).show()
+            }
+            mAlertDialog.setNegativeButton("No") { dialog, id ->
+                //perform som tasks here
+                Toast.makeText(context, "No", Toast.LENGTH_SHORT).show()
+            }
+            mAlertDialog.show()
+
+
 
         }
         //set the data as a class ContactDetails
